api/requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
bcrypt==4.1.2
cryptography==41.0.7

# API Documentation
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5

# Additional documentation tools
django-extensions==3.2.3

# File handling
Pillow==10.1.0
django-storages==1.14.2

# Development tools
django-extensions==3.2.3
django-debug-toolbar==4.2.0
ipython==8.17.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0

# Caching
redis==5.0.1
django-redis==5.4.0

# Task queue
celery==5.3.4
django-celery-beat==2.5.0

# Utils
python-slugify==8.0.1
requests==2.31.0
python-dateutil==2.8.2